+++ midgroupproject/app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Groupproject"
         tools:targetApi="31">
-        <activity android:name=".liked.LikedSongLibraryActivity"
+        <activity android:name=".playlist.LikedSongPlaylistActivity"
             android:exported="false" />
 
         <activity
