+++ midgroupproject/app/src/main/java/vn/edu/usth/midgroupproject/playlist/RecyclerView/SongModel.java
@@ -1,4 +1,4 @@
-package vn.edu.usth.midgroupproject.liked;
+package vn.edu.usth.midgroupproject.playlist.RecyclerView;
 
 public class SongModel {
     String songTitle;
