+++ midgroupproject/app/src/main/res/layout/activity_playlist.xml
@@ -6,22 +6,20 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/gradient_bg"
-    tools:context=".liked.LikedSongsPageFragment">
+    tools:context=".playlist.LikedSongPlaylistActivity">
 
     <LinearLayout
         android:id="@+id/linearLayout"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:orientation="vertical"
-        app:layout_constraintBottom_toTopOf="@id/bottomNavigationView"
         app:layout_constraintTop_toTopOf="parent">
 
         <LinearLayout
             android:id="@+id/main_frame"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:orientation="vertical"
-            app:layout_constraintEnd_toEndOf="parent">
+            android:orientation="vertical">
 
             <LinearLayout
                 android:layout_width="match_parent"
@@ -44,6 +42,7 @@
                 android:padding="15dp">
 
                 <TextView
+                    android:id="@+id/playlist_name"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:fontFamily="@font/spotify_mix"
@@ -51,13 +50,25 @@
                     android:textColor="@color/white"
                     android:textSize="20dp"
                     android:textStyle="bold" />
-
-                <TextView
-                    android:layout_width="match_parent"
+                <LinearLayout
+                    android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:fontFamily="@font/spotify_mix"
-                    android:text=". . .  songs"
-                    android:textColor="#B3B3B3" />
+                    android:orientation="horizontal">
+                    <TextView
+                        android:id="@+id/playlist_song_number"
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content"
+                        android:fontFamily="@font/spotify_mix"
+                        android:text=". . ."
+                        android:textColor="#B3B3B3" />
+                    <TextView
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content"
+                        android:fontFamily="@font/spotify_mix"
+                        android:text=" songs"
+                        android:textColor="#B3B3B3" />
+                </LinearLayout>
+
             </LinearLayout>
 
             <RelativeLayout
@@ -69,7 +80,7 @@
                 android:paddingBottom="1dp">
 
                 <ImageView
-                    android:id="@+id/download"
+                    android:id="@+id/btn_download"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_alignParentStart="true"
@@ -77,7 +88,7 @@
                     android:src="@drawable/down_arrow" />
 
                 <ImageView
-                    android:id="@+id/shuffle"
+                    android:id="@+id/btn_play_shuffle"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_alignParentEnd="true"
@@ -95,12 +106,12 @@
     </LinearLayout>
 
 
-    <com.google.android.material.bottomnavigation.BottomNavigationView
-        android:id="@+id/bottomNavigationView"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="#CC121212"
-        android:layout_alignParentBottom="true"
-        app:menu="@menu/navbar" />
+<!--    <com.google.android.material.bottomnavigation.BottomNavigationView-->
+<!--        android:id="@+id/bottomNavigationView"-->
+<!--        android:layout_width="match_parent"-->
+<!--        android:layout_height="wrap_content"-->
+<!--        android:background="#CC121212"-->
+<!--        android:layout_alignParentBottom="true"-->
+<!--        app:menu="@menu/navbar" />-->
 
 </RelativeLayout>
